Move Page component to correct folder structure
