+++ src/components/layout/layout.module.css
@@ -40,6 +40,22 @@
     border-color: #2970FF;
 }
 
+/* Page Body styles */
+
+.container {
+    background-color: #F3F5F6;
+}
+
+.pageMain {
+    max-width: 100%;
+    grid-row: auto !important;
+    grid-column: 4 / 12 !important;
+    margin-top: 6rem !important;
+    margin-bottom: 8rem !important;
+    margin-left: 0 !important;
+    padding: 0 0 0 0 !important;
+}
+
 /* Footer styles */
 
 .footer {
