--- src/pages/pages.module.css
@@ -1,13 +1,0 @@
-.container {
-    background-color: #F3F5F6;
-}
-
-.pageMain {
-    max-width: 100%;
-    grid-row: auto !important;
-    grid-column: 4 / 12 !important;
-    margin-top: 6rem !important;
-    margin-bottom: 8rem !important;
-    margin-left: 0 !important;
-    padding: 0 0 0 0 !important;
-}
